feat: put functions in different files
