+++ .gitattributes
@@ -1,1 +1,2 @@
 * text=auto
+* text eol=lf
