+++ lib/utils/clean_up.js
@@ -1,0 +1,8 @@
+
+const cleanUp = (str) => {
+	const cleanStr = str.replace(/^\s*[\r\n\n]/gmu, "\n")
+
+	return cleanStr
+}
+
+export default cleanUp
