+++ lib/utils/validators.js
@@ -1,0 +1,16 @@
+
+// Validate name
+const validateName = (path) => {
+	console.log(`Valid: ${path}`)
+
+	return true
+}
+
+// Validate path
+const validatePath = (path) => {
+	console.log(`Valid: ${path}`)
+
+	return true
+}
+
+export { validateName, validatePath }
